pyproject.toml
[project]
name = "gemini-didaktik"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "google-genai>=1.67.0",
    "python-dotenv>=1.2.2",
]
